Extract common beans to separate configuration
